[Refactoring]: Weather 관련 Code Refactoring 및 코드 설명 추가
- WeatherTestView 에서 Weather API 호출 후 어떤 값이 나오는지 테스트할 수 있습니다.

- 기존에 구현되어 있던 Weather Logic은 아직 합치지 않았습니다.
