+++ WeatherPlaylist/WeatherPlaylist/Source/Singleton/Weather/LocationManager.swift
