+++ WeatherPlaylist/WeatherPlaylist/Source/Singleton/Weather/WeatherLogic.swift
